backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
yfinance==0.2.37
fredapi==0.5.1
httpx==0.27.0
cachetools==5.3.3
